+++ src/components/file-card/FileCard.tsx
@@ -1,12 +1,16 @@
 import { PARAMETERS } from "@/config/parameters";
-import { AuthContext } from "@/context/AuthContext";
-import { UserFilesContext } from "@/context/UserFilesContext";
+import {
+  AuthContext,
+  FilesDialogsContext,
+  FoldersNavigationContext,
+  UserFilesContext
+} from "@/context/index";
 import { UserFilesActionTypes } from "@/hooks/user-files/UserFilesReducer";
 import { getFileByUUIDService } from "@/services/files/get-file-by-uuid.service";
 import { File } from "@/types/entities";
+import { NavigationParams } from "@/types/enums";
 import { FileText, FolderOpen } from "lucide-react";
 import { useContext, useEffect } from "react";
-import { useSearchParams } from "react-router-dom";
 import { toast } from "sonner";
 
 import { DropDown } from "./Dropdown";
@@ -36,15 +40,22 @@ const getFileIcon = (isFile: boolean) => {
 
 export const FileCard = ({ file }: { file: File }) => {
   const { session } = useContext(AuthContext);
+
+  const { pushToHistory } = useContext(FoldersNavigationContext);
+
   const { userFilesDispatcher } = useContext(UserFilesContext);
-  const [_params, setParams] = useSearchParams();
+  const { dialogsOpenState } = useContext(FilesDialogsContext);
 
   const downloadFile = () => {
     console.log("Downloading file");
   };
 
-  const navigateToFolder = () => {
-    setParams({ directory: file.uuid });
+  const handleDirectoryClick = () => {
+    if (dialogsOpenState.MOVE_FILE) {
+      pushToHistory(NavigationParams.MOVE_FILE, file.uuid);
+    } else {
+      pushToHistory(NavigationParams.DIRECTORY, file.uuid);
+    }
   };
 
   useEffect(() => {
@@ -57,7 +68,7 @@ export const FileCard = ({ file }: { file: File }) => {
     const checkIfFileIsReady = async (): Promise<boolean> => {
       const { success, ...res } = await getFileByUUIDService({
         fileUUID: file.uuid,
-        token: session?.token || ""
+        token: session?.token as string
       });
 
       if (!success || !res.file) {
@@ -111,7 +122,8 @@ export const FileCard = ({ file }: { file: File }) => {
     return (
       <button
         className="relative flex w-52 cursor-pointer flex-col items-center space-y-2 rounded-md border bg-primary-foreground/25 p-4 shadow-none transition-colors hover:bg-primary-foreground/75 hover:shadow-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-        onClick={file.isFile ? downloadFile : navigateToFolder}
+        onClick={file.isFile ? downloadFile : handleDirectoryClick}
+        aria-label={`${file.name} card`}
       >
         {<DropDown file={file} />}
         {getFileIcon(file.isFile)}