+++ src/screens/auth/profile/UpdatePasswordForm.tsx
@@ -8,7 +8,7 @@ import {
   FormMessage
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { AuthContext } from "@/context/AuthContext";
+import { AuthContext } from "@/context/index";
 import { updatePasswordService } from "@/services/account/update-password.service";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Loader2 } from "lucide-react";
@@ -49,7 +49,7 @@ export const UpdatePasswordForm = () => {
     const { success, msg } = await updatePasswordService({
       oldPassword: values.oldPassword,
       newPassword: values.newPassword,
-      token: session?.token || ""
+      token: session?.token as string
     });
     if (!success) {
       setLoading(false);