+++ src/types/enums.ts
@@ -1,0 +1,10 @@
+export enum Dialogs {
+  RENAME_FILE = "RENAME_FILE",
+  MOVE_FILE = "MOVE_FILE",
+  CREATE_DIRECTORY = "CREATE_DIRECTORY"
+}
+
+export enum NavigationParams {
+  DIRECTORY = "directory",
+  MOVE_FILE = "moveTo"
+}