+++ CHANGELOG.md
@@ -1,27 +1,45 @@
-# [0.3.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.2.0...v0.3.0) (2023-10-17)
+# [0.7.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.6.1...v0.7.0) (2023-10-20)
 
 
 ### Features
 
-* Profile page and update password ([#16](https://github.com/PedroChaparro/unofficial-frontend-react/issues/16)) ([2b8825f](https://github.com/PedroChaparro/unofficial-frontend-react/commit/2b8825ff55b8df092b72bbaaab9a9544b1e1307b))
+* Create folder ([#24](https://github.com/PedroChaparro/unofficial-frontend-react/issues/24)) ([54dc7c1](https://github.com/PedroChaparro/unofficial-frontend-react/commit/54dc7c15c36fa3b19bad68caed62a6c45e8ccb54))
 
 
 
-# [0.2.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.1.0...v0.2.0) (2023-10-16)
+## [0.6.1](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.6.0...v0.6.1) (2023-10-20)
+
+
+### Bug Fixes
+
+* Go back button in folders navigation ([#23](https://github.com/PedroChaparro/unofficial-frontend-react/issues/23)) ([07edb44](https://github.com/PedroChaparro/unofficial-frontend-react/commit/07edb442b2524bd739bf7f0d6ea13f536b1a62b9))
+
+
+
+# [0.6.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.5.0...v0.6.0) (2023-10-20)
 
 
 ### Features
 
-* Dialog to upload files ([#15](https://github.com/PedroChaparro/unofficial-frontend-react/issues/15)) ([1441752](https://github.com/PedroChaparro/unofficial-frontend-react/commit/14417526a6eeb4fc3caaab4f6b37a53e65b35052))
+* Move files ([#21](https://github.com/PedroChaparro/unofficial-frontend-react/issues/21)) ([3a470e9](https://github.com/PedroChaparro/unofficial-frontend-react/commit/3a470e97df16f038658bb2e13019c2adda4a5e0a))
 
 
 
-# [0.1.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/410290a1e5a90644d7e76d422f5504a0aff4fbe2...v0.1.0) (2023-10-16)
+# [0.5.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.4.0...v0.5.0) (2023-10-20)
 
 
 ### Features
 
-* List user files ([#14](https://github.com/PedroChaparro/unofficial-frontend-react/issues/14)) ([410290a](https://github.com/PedroChaparro/unofficial-frontend-react/commit/410290a1e5a90644d7e76d422f5504a0aff4fbe2))
+* Home page ([#20](https://github.com/PedroChaparro/unofficial-frontend-react/issues/20)) ([36a272a](https://github.com/PedroChaparro/unofficial-frontend-react/commit/36a272ad72732e2eb8b21c8a18a9dfab7d72c837))
+
+
+
+# [0.4.0](https://github.com/PedroChaparro/unofficial-frontend-react/compare/v0.3.0...v0.4.0) (2023-10-17)
+
+
+### Features
+
+* Rename files ([#18](https://github.com/PedroChaparro/unofficial-frontend-react/issues/18)) ([bbc6f61](https://github.com/PedroChaparro/unofficial-frontend-react/commit/bbc6f611b9f8dd9526d67e55bc79aabc1a9acdb2))